refactor - enemyAI 패키지 이동
